Moved authentication to the configuration extensions
